requirements.txt
Django
pymongo
bcrypt
langchain
langchain-community
sentence-transformers
python-dotenv
pinecone-client
huggingface-hub